requirements.txt
biopython==1.85
contourpy==1.3.2
cycler==0.12.1
fair-esm==2.0.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
